Backport - Make product thumbnail configurable
As it is now the product thumbnail image is hardcoded to 75x75 so if you want to display larger images on the autocomplete list the only choice is to use CSS but this will make images look blurry.

It would make sense to use the `product_thumbnail_image` image size here instead which is already 75x75 by default and, if anyone needs a different custom size, it can be overwritten using the `etc/view.xml` file.